WebwithMCP-main/requirements.txt
aiohttp==3.12.15
aiomysql==0.2.0
aiosqlite==0.21.0
fastapi==0.116.1
httpx==0.28.1
langchain_core==0.3.74
langchain_mcp_adapters==0.1.9
langchain_openai==0.3.31
pydantic>=2.0,<3.0
python-dotenv==1.1.1
uvicorn==0.35.0